modify: make components
