+++ lib/game_components/game_components.dart
@@ -1,0 +1,2 @@
+export 'base_game.dart';
+export 'life.dart';
