+++ lib/logic/entities/entities.dart
@@ -1,0 +1,1 @@
+export 'coordinate.dart';
